[MBV] Replace NOP4 with MERKLEBRANCHVERIFY in witness scripts.
  [... leaf/H(leaf) ...] proof root {count,prehashed} MERKLEBRANCHVERIFY

This opcode takes a Merkle root, any number of leaf hashes of the tree, and a 'proof' object composed of an encoding of the path through the tree from root to leaves and the minimal set of hash values necessary to recompute the the root from the leaf. Script validation fails if the recomputed root does not match the hash provided (which, presumably, would be committed to in the scriptPubKey).

See Bitcoin's BIP116 for details. Unlike BIP116, this implementation drops its first 3 arguments.