[Demurrage] Add 3.2% demurrage using an ALU implementation of TimeAdjustValueForward.
Changes TimeAdjustValueForward to use an integer-math-only method for performing demurrage calculations.  Note that due to truncation of the intermediate multiplication terms it is possible for this method to disagree with the results of the previously used MPFR calculation by as much as 1 kria, even after the latter is truncated.  It boasts a 10x improvement in speed, however.