Revert "qt: Remove thousands separators after decimal point"
This reverts commit a95b1199dbcf6c90508250d9b4cfd1df78106a5b.

The SI standard describes digit-separating thin spaces as an optional visual cue, but it says nothing of using thin spaces for the whole part of the number only, and no spaces for the decimal part.  This is an arbitrary design decision based on aesthetics.  This conform us with international standards by reverting the removal of thin-space digit separators for the decimal portion of displayed currency amounts.