[Soft-fork] Require the nTimeLock value of a coinbase transaction to be equal to the median of the past 11 block times.
This is an old-style supermajority rollout with BIP8/BIP9-like version bits and BIP8-like activation-on-timeout semantics.  It's not strictly compatible with those BIPs though, as we didn't want to back port all the necessary state management code.  But miners using BIP8/BIP9 compatible version bit software were able to signal for this rollout by configuring their devices to signal bit #28 (the highest-order BIP8/BIP9 bit).  Bit 28 needs to be set during the activation period between April 2, 2019 and October 2, 2019 for a block to be considered valid.  This is residual condition from a pre-versionbits soft-fork deployment mechanism, and should not trigger the normal warnings about unexpected versionbits deployments.