[Wallet] Correctly apply demurrage to watch-only and immature balance calculations.
Version 10.4 introduced the new feature of watch-only addresses,
which are scripts the wallet tracks as if it had the private keys,
a construct which finds use in many circumstances such as online
tracking of a cold-storage wallet. Unfortunately the wallet routines
for calculating watch-only balances were not updated to apply
demurrage to these amounts. As a result, the display for a watch-only
wallet shows an incorrectly inflated amount of total freicoin compared
to what is actually available.

While fixing this bug it was re-discovered that immature balances were
also being displayed improperly, the result of a conscious decision in
a prior release for reasons that no longer apply.

This commit fixes both issues, and the GUI and RPC commands which
calculate wallet balances now correctly show demurrage-adjusted
totals.