[Segwit] Mask Coinbase's nSequence out of witness commitment.
The Stratum mining protocol requires some number of bytes of the Coinbase, typically 4 bytes, to be under the miner's control.  And existing mining software is unaware of block-final transactions or our witness commitment mechanism.  To maintain compatibility with this critical, already deployed infrastructure, both the scriptSig and nSequence fields of the coinbase transaction are truncated/zero'd when calculating the witness commitment hash.  This allows any of these 4 bytes to be set by the miner without altering the witness hash, which would require re-calculating the transaction Merkle tree.