[Merkle] Add stable Merkle branch proof APIs.
The Satoshi Merkle tree design will duplicate hash values along the right-most branch of the tree if it is not a power of 2 in size.  In the regular Merkle branch APIs, these hashes are included in the branch proof.  This not only wastes space, but is problematic because these duplicated hashes are dependent on the leaf value being proven, so the proof can't be used to recalculate a new root if the leaf value changes.

This new API for generating Merkle tree branch proofs does not include duplicated hashes, so the result is both shorter (if it is along the right-hand side of an unbalanced tree) and can be safely used to recalculate root hash values.