[Warn] Do not allow the user to start without `-connect` to trusted peer.
The v12 branch does not distinguish between the protocol-cleanup rule changes and the relaxation of aggregate block limits in the size-expansion flag day.  Since v12 branch clients also do not support the auxiliary proof-of-work rules either, it is trivial for a malicious peer to perform a disk-space exhausting denial of service attack against v12 peers once the protocol-cleanup rule change activates on April 16th, 2021.

To prevent such attacks, we refuse to start up unless the user has configured a trusted node to connect to with the `-connect` parameter.