[FinalTx] Force -reindex on upgrade to for block-undo format change.
While the utxo-set database can be upgraded in-place using upstream code from Bitcoin, the block-final transaction record cannot, as it depends on data that is not presently contained in the utxo-set, and requires rewriting the undo files.  Upon first launch the user will get a notification that they need to reindex, which can be done automatically in the GUI or by providing the '-reindex' option on the command line.