[IsStandard] Lower minimum (stripped) transaction size to 65 bytes.
It is possible to create small transactions when spending outputs that do not require legacy signatures (including native segwit) and sending to OP_TRUE outputs, as some RPC tests do.  Due to a weakness in the block-transaction Merkle tree (CVE-2012-2459), allowing a 64-byte transaction potentially enables commitment to multiple transactions using a partial-hash collision.  To make pulling off such an attack more challenging, we keep the minimum transaction size but set it to 65 bytes.