[Sign] Remove SIGHASH_ALL as an API-level default from various signing fuctions.
Since refheights and SIGHASH modes are both ints, the default parameter behavior of the signing mode can potentially cause bugs when a SIGHASH mode is mistakenly processed as a refheight parameter.  To avoid this, we remove the default setting of SIGHASH mode to SIGHASH_ALL, so any un-updated calls to these signing functions will result in a compiler error (too few arguments provided).