[Mining] Allow rpc.getblocktemplate() to be called without network connections on regtest.
Previously a JSONRPCError would be thrown if getblocktemplate is called on a node that is not connected to the network.  This is a good safeguard for public networks, but is just an annoying restriction to work around in test code.  This commit removes that restriction for regtest, or any other network that allows mining of blocks on demand.