[SizeExpansion] Remove MaxProtocolMessageLength calculation for 32-bit nodes.
MaxProtocolMessageLength returns a nifty calculated value for 32-bit nodes based on the maximum number of allowed network connections.  This allows 32-bit nodes which would otherwise be RAM-limited to scale with the network block size by limiting their p2p connections.  Unfortunately something about this breaks the gitian compilation environment in a way that is not well characterized yet.

This commit removes that calculation logic and unifies behavior on both 32-bit and 64-bit hosts.  32-bit hosts which want to protect themselves from memory exhaustion DoS attacks after the activation of the size expansion rule change must either stay on v18 or find a way to upgrade to a 64-bit client.