[Refheight] Separate "value" from "amount" in RPC APIs.
In Bitcoin the terms "value" and "amount" are largely interchangeable in meaning.  In Freicoin they adopt very specific, technical meaning.  A "value" is a raw output nValue field as seen on the block chain.  An "amount" is the time-adjusted value of that txout when it serves as input into another transaction.  This commit renames JSON-RPC fields which refer to the txout.nValue field as "amount" to "value".