[Locale] Remove all bitcoin translation files except the original English source.
The English translation files are mechanically transformed by the rebranding scripts, however there is no guarantee that the process works correctly for non-English languages.  We remove all bitcoin translation files and then later add reviewed and edited translation sources for the languages that we support.  Since at least one translation is required by the build scripts on some platforms, we create an "en_US" translation which is a copy of the source.