[Demurrage] Bias GetTimeAdjustedValue prior to BIP66 activation, to compensate for truncation.
Following the switch to integer-math-only demurrage calculations, every call to GetTimeAdjustedValue() in the validation of the historical blockchain is potentially off-by-1.  To prevennt rejection of historical transactions, we add one kria as input for each of the calls that would have been made, in every transaction prior to activation of the new integer math demurrage calculation rules.